requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
colorama==0.4.6
cssselect==1.2.0
distlib==0.3.9
dnspython==2.7.0
filelock==3.16.1
fsspec==2024.10.0
huggingface-hub==0.26.3
identify==2.6.3
idna==3.10
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.1.3
packaging==24.2
platformdirs==4.3.6
praw==7.8.1
prawcore==2.4.0
pre_commit==4.0.1
pymongo==4.10.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
setuptools==75.6.0
soupsieve==2.6
sympy==1.13.1
tokenizers==0.20.3
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
update-checker==0.18.0
urllib3==2.2.3
vaderSentiment==3.3.2
virtualenv==20.28.0
websocket-client==1.8.0